% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cdogr_codon.R
\name{Cdogr_codon}
\alias{Cdogr_codon}
\title{Title}
\usage{
Cdogr_codon(dna_data, start = 1)
}
\arguments{
\item{dna_data}{Insert proper DNA string}

\item{start}{Pick start codon position}
}
\value{
Returns the corresponding codons for the DNA string
}
\description{
Title
}
