% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cdogr_frequencies.R
\name{Cdogr_frequencies}
\alias{Cdogr_frequencies}
\title{Visualizing Amino Acid Frequencies}
\usage{
Cdogr_frequencies(frq2)
}
\arguments{
\item{frq2}{..}
}
\value{
Amino Acid Frequencies
}
\description{
Visualizing Amino Acid Frequencies
}
\examples{
..
}
