% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cdogr_splitting.R
\name{Cdogr_splitting}
\alias{Cdogr_splitting}
\title{Splitting RNA into Codons}
\usage{
Cdogr_splitting(splt2, start = 1)
}
\arguments{
\item{splt2}{RNA}

\item{start}{dfv}
}
\value{
Returns codons
}
\description{
Splitting RNA into Codons
}
