% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cdogr_substitution.R
\name{Cdogr_substitution}
\alias{Cdogr_substitution}
\title{Converting DNA to RNA}
\usage{
Cdogr_substitution(subT2)
}
\arguments{
\item{subT2}{dfvx}
}
\value{
fvdfv
}
\description{
Converting DNA to RNA
}
\examples{
vf
}
