% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Codon_table}
\alias{Codon_table}
\title{Codon table}
\format{
An object of class \code{data.frame} with 1 rows and 64 columns.
}
\source{
\url{https://www.ncbi.nlm.nih.gov/Taxonomy/Utils/wprintgc.cgi?chapter=tgencodes#SG1}
}
\usage{
Codon_table
}
\description{
The table gives us an overview of all the codons, and the amino acid residues that match the codons.
}
\keyword{datasets}
