% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cdogr_translating.R
\name{translating}
\alias{translating}
\title{Translating RNA codons into amino acids}
\usage{
translating(codons)
}
\arguments{
\item{codons}{}
}
\value{
Amino Acid Residues
}
\description{
Translating RNA codons into amino acids
}
\examples{
..
}
